requirements.txt
fabric
django
django-debug-toolbar
psycopg2
python-memcached
south
django-taggit
-e hg+https://bitbucket.org/fabian/django-taggit-autosuggest#egg=django-taggit-autosuggest
django-ratings
django-tinymce